+++ CHANGELOG.md
@@ -16,6 +16,7 @@ Translations: [简体中文](CHANGELOG.zh.md)
   ImageInfo. [#259](https://github.com/panpf/sketch/issues/259)
 * fix: Fixed a bug where MemoryCacheRequestInterceptor failed to load images when they encountered a
   cache without ImageInfo. [#260](https://github.com/panpf/sketch/issues/260)
+* new: MemoryCache added entries() method to get all cache entries
 
 # 4.2.0